powo-builder/requirements.txt
google-cloud-dns==0.29.0
kubernetes==6.0.0
requests==2.19.1
